Fix GCC 4.0.3 warning - size_t cannot be negative
ServerMenu.cxx:181: warning: converting negative value ‘-0x00000000000000001’ to ‘size_t’

git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18545 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/HUDuiTabbedControl.h
@@ -33,8 +33,8 @@ class HUDuiTabbedControl : public HUDuiNestedContainer {
     std::string getActiveTabName() { return tabs[activeTab].first; }
 
     void setActiveTab(size_t tab);
-	
-    void addTab(HUDuiControl* tabControl, std::string tabName, size_t index = -1);
+
+    void addTab(HUDuiControl* tabControl, std::string tabName, size_t index = (size_t) -1);
     void removeTab(HUDuiControl* control, std::string tabName);
     void removeTab(size_t tabIndex);
